Tweak GTP wording, document and tweak hotbar mishap
